Implement UPDATE and GET methods on admin load endpoint (#26)
Implement UPDATE and GET methods on admin load endpoint

+++ src/database/append_smiley.rs
@@ -11,8 +11,8 @@ pub struct RestaurantWithSmileyReport {
     pub city: String,
     pub cvr: String,
     pub pnr: String,
-    pub latitude: f32,
-    pub longitude: f32,
+    pub latitude: f64,
+    pub longitude: f64,
     pub version_number: i32,
     pub smileyreports: Vec<SmileyReport>,
 }
